final configs
