+++ utils/benchmark_rocks_config/configs/config_16_Final_02.rs
@@ -1,7 +1,6 @@
 use rocksdb::BlockBasedOptions;
 use rocksdb::Cache;
 use rocksdb::Options;
-use rocksdb::PlainTableFactoryOptions;
 
 pub enum CacheSetting {
     /// Enabled cache with the given size in bytes
@@ -22,20 +21,17 @@ impl Default for DbConfig {
 }
 
 impl DbConfig {
-    pub fn to_options(self, cache_setting: CacheSetting, prefix_len: Option<usize>, key_len: usize) -> Options {
+    pub fn to_options(self, cache_setting: CacheSetting, prefix_len: Option<usize>, _key_len: usize) -> Options {
         let mut opts = Options::default();
         let mut block_based_options = BlockBasedOptions::default();
 
         opts.create_if_missing(true);
         opts.create_missing_column_families(true);
         opts.increase_parallelism(16);
-        // add an assertion so that the ulimit -n is at least double this.
-        // opts.set_max_open_files(40960);
 
         block_based_options.set_pin_l0_filter_and_index_blocks_in_cache(true);
         block_based_options.set_cache_index_and_filter_blocks(true);
         block_based_options.set_bloom_filter(15.5, true);
-        opts.set_allow_concurrent_memtable_write(false);
 
         // NOTE: As per the rocks db wiki: "The overhead of statistics is usually small but non-negligible. We usually observe an overhead of 5%-10%."
         #[cfg(feature = "metrics")]
@@ -47,8 +43,7 @@ impl DbConfig {
         if let Some(prefix_len) = prefix_len {
             let transform = rocksdb::SliceTransform::create_fixed_prefix(prefix_len);
             block_based_options.set_index_type(rocksdb::BlockBasedIndexType::HashSearch);
-            opts.set_memtable_whole_key_filtering(true);
-            opts.set_memtable_prefix_bloom_ratio(0.2);
+            opts.set_memtable_prefix_bloom_ratio(0.15);
             opts.set_prefix_extractor(transform);
         }
 
@@ -62,22 +57,12 @@ impl DbConfig {
 
         match self {
             DbConfig::OptimizedPointLookUp => {
-                let plain = PlainTableFactoryOptions {
-                     user_key_length: key_len as u32,
-                     bloom_bits_per_key: 15,
-                     hash_table_ratio: 0.75,
-                     index_sparseness: 16,
-                     huge_page_tlb_size: 0,
-                     encoding_type: rocksdb::KeyEncodingType::Plain,
-                     full_scan_mode: false,
-                     store_index_in_file: true
-                 };
+                block_based_options.set_data_block_hash_ratio(0.3);
+                block_based_options.set_data_block_index_type(rocksdb::DataBlockIndexType::BinaryAndHash);
 
                 opts.set_use_direct_reads(true);
                 opts.set_memtable_whole_key_filtering(true);
-
                 opts.set_compression_type(rocksdb::DBCompressionType::None);
-                opts.set_plain_table_factory(&plain);
             }
             DbConfig::Default => {
                 opts.set_compression_type(rocksdb::DBCompressionType::Lz4);
