+++ utils/benchmark_rocks_config/configs/config_17_Final_03.rs
@@ -31,7 +31,7 @@ impl DbConfig {
 
         block_based_options.set_pin_l0_filter_and_index_blocks_in_cache(true);
         block_based_options.set_cache_index_and_filter_blocks(true);
-        block_based_options.set_bloom_filter(15.5, true);
+        block_based_options.set_bloom_filter(15.5, false);
 
         // NOTE: As per the rocks db wiki: "The overhead of statistics is usually small but non-negligible. We usually observe an overhead of 5%-10%."
         #[cfg(feature = "metrics")]
@@ -42,19 +42,25 @@ impl DbConfig {
 
         if let Some(prefix_len) = prefix_len {
             let transform = rocksdb::SliceTransform::create_fixed_prefix(prefix_len);
-            opts.set_memtable_prefix_bloom_ratio(0.02);
+            block_based_options.set_index_type(rocksdb::BlockBasedIndexType::HashSearch);
+            opts.set_memtable_prefix_bloom_ratio(0.15);
             opts.set_prefix_extractor(transform);
         }
 
         if let CacheSetting::Enabled(cache_size) = cache_setting {
-            let cache = Cache::new_lru_cache(cache_size);
-            block_based_options.set_block_cache(&cache);
+            let block_cache = Cache::new_lru_cache(cache_size/2);
+            let row_cache = Cache::new_lru_cache(cache_size/2);
+
+            opts.set_row_cache(&row_cache);
+            block_based_options.set_block_cache(&block_cache);
         }
 
         match self {
             DbConfig::OptimizedPointLookUp => {
-                block_based_options.set_data_block_hash_ratio(0.5);
+                block_based_options.set_data_block_hash_ratio(0.3);
                 block_based_options.set_data_block_index_type(rocksdb::DataBlockIndexType::BinaryAndHash);
+
+                opts.set_use_direct_reads(true);
                 opts.set_memtable_whole_key_filtering(true);
                 opts.set_compression_type(rocksdb::DBCompressionType::None);
             }
@@ -67,6 +73,7 @@ impl DbConfig {
         }
 
         opts.set_block_based_table_factory(&block_based_options);
+
         opts
     }
 }
